trigger_var_control/trigger_var_control/requirements.txt
numpy
pandas
openpyxl
cimpy
flask
requests
villas-dataprocessing
OpenDSSDirect.py==0.9.4